src/requirements.txt
cssselect==1.0.3
cycler==0.10.0
dukpy==0.2.2
elasticsearch==6.3.1
future==0.17.1
javascripthon==0.10
Jinja2==2.10
jupyter-echarts-pypkg==0.1.2
kiwisolver==1.0.1
lml==0.0.2
lxml==4.3.2
macropy3==1.1.0b2
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.16.2
Pillow==5.4.1
pkuseg==0.0.18
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.1.10
PyMySQL==0.9.3
pyparsing==2.3.1
pyquery==1.4.0
python-dateutil==2.8.0
PyYAML==3.13
six==1.12.0
urllib3==1.24.1
wordcloud==1.5.0
